refactor: give selectedGeneratedDisplayPair a single home
The function was byte-identical in encrypt-page.tsx and key-detail-dialog.tsx.
